// backend/src/modules/dinosaurs/models/epoch.enum.ts
export enum Epoch {
    Prehistoric_Epoch1 = 'Prehistoric_Epoch1',
    Prehistoric_Epoch2 = 'Prehistoric_Epoch2',
    Prehistoric_Epoch3 = 'Prehistoric_Epoch3',
    Prehistoric_Epoch4 = 'Prehistoric_Epoch4',
    Ancient_Epoch1 = 'Ancient_Epoch1',
    Ancient_Epoch2 = 'Ancient_Epoch2',
    Ancient_Epoch3 = 'Ancient_Epoch3',
    Ancient_Epoch4 = 'Ancient_Epoch4',
    Medieval_Epoch1 = 'Medieval_Epoch1',
    Medieval_Epoch2 = 'Medieval_Epoch2',
    Medieval_Epoch3 = 'Medieval_Epoch3',
    Medieval_Epoch4 = 'Medieval_Epoch4',
    Modern_Epoch1 = 'Modern_Epoch1',
    Modern_Epoch2 = 'Modern_Epoch2',
    Modern_Epoch3 = 'Modern_Epoch3',
    Modern_Epoch4 = 'Modern_Epoch4',
    Future_Epoch1 = 'Future_Epoch1',
    Future_Epoch2 = 'Future_Epoch2',
    Future_Epoch3 = 'Future_Epoch3',
    Future_Epoch4 = 'Future_Epoch4',
  }